Consolidate extraction of course colours
