+++ apps/website/src/components/lander/course-content/FutureOfAiContent.tsx
@@ -6,21 +6,9 @@ import {
   PiUsers,
 } from 'react-icons/pi';
 import { type CourseLanderContent } from '../CourseLander';
+import { FOAI_COLORS } from '../../../lib/courseColors';
 
 export const FUTURE_OF_AI_START_URL = '/courses/future-of-ai/1/1';
-
-export const FOAI_COLORS = {
-  /* Gradient layers:
-     1. Left-side darkness - horizontal gradient to darken left edge
-     2. Top-right subtle glow - very faint greenish highlight
-     3. Colorful gradient - green→gold→purple from top-right (reduced opacity for muted look)
-     4. Base color - dark olive */
-  gradient: 'linear-gradient(to right, rgba(30, 30, 20, 0.6) 0%, rgba(30, 30, 20, 0.4) 25%, rgba(30, 30, 20, 0.2) 45%, transparent 60%), radial-gradient(ellipse 70% 60% at 85% 20%, rgba(155, 180, 115, 0.12) 0%, transparent 60%), radial-gradient(ellipse 200% 180% at 105% -5%, rgba(150, 207, 156, 0.35) 0%, rgba(163, 179, 110, 0.35) 28.6%, rgba(176, 152, 64, 0.35) 57.2%, rgba(147, 120, 64, 0.35) 67.9%, rgba(118, 88, 64, 0.35) 78.6%, rgba(89, 56, 63, 0.35) 89.3%, rgba(60, 24, 63, 0.35) 100%), #29281D',
-  accent: '#E6DBA6',
-  iconBackground: '#64663E',
-  bright: '#faf6e1',
-  full: '#8c8146',
-};
 
 export const createFutureOfAiContent = (
   _applicationUrlWithUtm: string,
